Selecting and Manipulating Elements
Document Object Model is structed representation of html documents. Allow Js to access html elements ans styles to manipulate them. DOM is connection point between js and html. Also like document.querySelector document is an special object as entry point to the DOM. DOM is not part of JS. It's part of Web APIs also was written with js.
For Input element's content we use value, textContent for other elemet's