Scoping in Practice -- Part 3
If the variable was created in function scope with same name an another value had created before in parents scope, that is not a problem.  Actually this is how it works on same parameter name in functions. Because they have their own execution context. So the variable is completely created as a new variable.