Short Circuiting
the OR operator will return the first  truthy value of all the operands or simply the last value if all of them are falsy. The And operator will return the first falsy value or the last value if all of them are truthy. We can use OR operator to ser default values and we can use the AND operator to execute code in the second operand