Working With Strings --Part I
str.slice() first index is included to calculation but last index don't use last index-1 is used for slice the string. We use metods of String object in Javascript actually .  At the behind of JS  while we using string methods string is converted to String Object and its methods can be used. And of the calculation Js convert the result string again.